requirements.txt
celery==5.1.2
discord==1.7.3
dateparser==1.0.0
django-autoslug==1.9.8
django-cors-headers==3.7.0
django-datatables-view==1.18.0
django-simple-history==3.0.0
Django==3.2.5
djangorestframework-gis==0.17
djangorestframework==3.12.4
Faker==4.0.0
gunicorn==20.1.0
pillow==6.1.0
psycopg2==2.8.5
redis==3.5.2
requests==2.24.0
sendgrid==6.4.8
xlrd==2.0.1
xlwt==1.3.0